Reorganized test package into unit_test and integration_test subpackages.
Modified deploy_tools to run script installing dependencies if pip.__version__
>= 10.0.0
