+++ mycity/mycity/deploy_tools/deploy_tools.py
@@ -7,6 +7,7 @@ import argparse
 import os
 import pip
 import shutil
+import subprocess
 import zipfile
 
 # TODO: This needs to be updated to give user some idea of how to run it and where to run it from
@@ -15,6 +16,8 @@ TEMP_DIR_PATH = os.path.join(PROJECT_ROOT, 'temp')
 LAMBDA_REL_PATH = 'platforms/amazon/lambda/custom/lambda_function.py'
 LAMBDA_FUNCTION_PATH = os.path.join(PROJECT_ROOT, LAMBDA_REL_PATH)
 MYCITY_PATH = os.path.join(PROJECT_ROOT, 'mycity')
+
+INSTALL_REQUIREMENTS_SCRIPT = os.path.join(os.getcwd(), 'install_requirements.sh')
 
 
 def zip_lambda_function_directory(zip_target_dir):
@@ -42,6 +45,22 @@ def zip_lambda_function_directory(zip_target_dir):
 
 def install_pip_dependencies(requirements_path, requirements_path_no_deps):
     """
+    Wraps both _install_pip_dependencies. pip version is checked in this
+    function so correct install function is called.
+
+    if pip version >= '10.0.0', we should install from script
+    else, install using pip.main
+    """
+    version = pip.__version__
+    if version >= '10.0.0':
+        _install_pip_dependencies_from_script()
+    else:
+        _install_pip_dependencies(requirements_path, requirements_path_no_deps)
+
+
+
+def _install_pip_dependencies(requirements_path, requirements_path_no_deps):
+    """
     Uses requirements.txt to install all external libraries used by the project
     in the temporary directory the .zip file is created from.
 
@@ -49,12 +68,23 @@ def install_pip_dependencies(requirements_path, requirements_path_no_deps):
     :return: none
     """
     print('Installing dependencies ... ', end='')
-    install_args = ["install", "-r", requirements_path, "-t", TEMP_DIR_PATH]
+    install_args = "pip install -r " + requirements_path + " -t " + TEMP_DIR_PATH
     pip.main(install_args)
     print('Installing dependencies from requirements_no_deps.txt ...', end='')
-    install_args = ["install", "--no-deps", "-r", requirements_path_no_deps,
-                    "-t", TEMP_DIR_PATH]
-    pip.main(install_args)
+    install_args_no_deps = ("pip install --no-deps -r ",
+                            requirements_paths_no_deps,
+                            " -t ", 
+                            TEMP_DIR_PATH)
+    pip.main(install_args_no_deps)
+    print('DONE')
+
+def _install_pip_dependencies_from_script():
+    """
+    pip.main is deprecated in the latest version of pip. If pip is version 10.0.0 
+    or later, we can install our dependencies using a shell script
+    """
+    print('Installing dependencies ... ', end='')
+    subprocess.call(INSTALL_REQUIREMENTS_SCRIPT)
     print('DONE')
 
 
@@ -69,7 +99,7 @@ def package_lambda_function():
     print('Creating temporary build directory ... ', end='')
     # create the temporary directory for the zip file's contents
     os.mkdir(TEMP_DIR_PATH)
-    # copy the lambda file, mycity files, andrequirements.txt to temp directory
+    # copy the lambda file, mycity files, and requirements.txt to temp directory
     shutil.copy(LAMBDA_FUNCTION_PATH, TEMP_DIR_PATH)
     shutil.copytree(MYCITY_PATH, os.path.join(TEMP_DIR_PATH, 'mycity'))
     requirements_path = os.path.join(os.getcwd(), 'requirements.txt')
