+++ mycity/mycity/test/integration_tests/README
@@ -1,0 +1,1 @@
+Smoke tests for intents, with web resources mocked
