+++ mycity/run_tests.py
@@ -1,11 +1,11 @@
 import unittest
 
-import mycity.test.test_location_utils as loc_utils
-import mycity.test.test_snow_emergency_parking as snow_parking
-import mycity.test.test_trash_intent as trash_intent
-import mycity.test.test_unhandled_intent as unhandled_intent
-import mycity.test.test_get_alerts as get_alerts
-import mycity.test.test_mycity_controller as my_controller
+import mycity.test.integration_tests.test_snow_emergency_parking as snow_parking
+import mycity.test.integration_tests.test_trash_intent as trash_intent
+import mycity.test.integration_tests.test_unhandled_intent as unhandled_intent
+import mycity.test.integration_tests.test_get_alerts as get_alerts
+import mycity.test.unit_tests.test_location_utils as loc_utils
+import mycity.test.unit_tests.test_mycity_controller as my_controller
 
 
 TEST_CASES = [ my_controller.MyCityControllerUnitTestCase,
